src/hooks/index.ts
// Hook exports
export { default as useAsync } from './useAsync';
export { default as useClickOutside } from './useClickOutside';
export { default as useDebounce } from './useDebounce';
export { default as useIntersectionObserver } from './useIntersectionObserver';
export { default as useKeyPress } from './useKeyPress';
export { default as useLocalStorage } from './useLocalStorage';
export { default as usePrevious } from './usePrevious';
export { default as useThrottle } from './useThrottle';

// Type exports
export type {
  UseAsyncOptions,
  UseAsyncResult,
} from './useAsync';

export type {
  UseDebounceOptions,
} from './useDebounce';

export type {
  UseIntersectionObserverOptions,
  UseIntersectionObserverResult,
} from './useIntersectionObserver';

export type {
  UseKeyPressOptions,
} from './useKeyPress';

// Utility function exports
export {
  isEnter,
  isEscape,
  isSpace,
  isArrowUp,
  isArrowDown,
  isArrowLeft,
  isArrowRight,
  isTab,
  isDelete,
  isBackspace,
  withCtrl,
  withShift,
  withAlt,
} from './useKeyPress';

// Re-export commonly used types from React
export type {
  RefObject,
  MutableRefObject,
  Dispatch,
  SetStateAction,
} from 'react';
